requirements.txt
beautifulsoup4==4.7.1
certifi==2019.6.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cssselect==1.0.3
cssutils==1.0.2
Django==2.1.1
django-cors-headers==3.0.2
django-filter==2.1.0
django-htmlmin==0.9.1
django-redis==4.10.0
django-tinymce==2.8.0
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
drf-extensions==0.5.0
emoji==0.5.2
html5lib==1.0.1
idna==2.8
itypes==1.1.0
Jinja2==2.10.1
jsonfield==2.0.2
lxml==4.3.4
Markdown==3.1.1
MarkupSafe==1.1.1
nap==2.0.0
Pillow==6.1.0
premailer==3.0.1
PyJWT==1.7.1
PyMySQL==0.9.3
pytz==2019.1
redis==3.2.1
requests==2.9.1
six==1.12.0
soupsieve==1.9.2
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.3
webencodings==0.5.1
